--improve calc for Multivariate Hypergeometric dist
This calc is designed to provide number of training examples required to have a certain minimum number of object-based examples present in the training set.  With higher dimensional objects, such as spheres, there are vastly more non-truth examples (that do not correspond to an underlying sphere) than those that do. We need to determine how many training examples are needed to be reasonably sure we have multiple representations of underlying objects in the training set.